Fix undefined behaviour in generateKey() and improve spread
The shift operation promotes the byte value to an int, however that's
only 32 bits and potentially -ve if char is signed so we end up
shifting more than the size of the type or shifting a -ve number. Both
of these are UB so add a cast to make sure that we operate on the
correct type.

+++ plugins/videoshape/VideoData.cpp
@@ -322,10 +322,10 @@ void VideoData::copyToTemporary(QIODevice &device)
 
 qint64 VideoData::generateKey(const QByteArray &bytes)
 {
-    qint64 answer = 1;
+    quint64 answer = 1;
     const int max = qMin(8, bytes.count());
     for (int x = 0; x < max; ++x)
-        answer += bytes[x] << (8 * x);
+        answer += static_cast<quint64>(bytes[x]) << (8 * x);
     return answer;
 }
